lsp: Move some logic into `SphinxClient`
This commit moves some of the details of starting the sphinx agent and
creating the application object into the `SphinxClient`.

It also records the response to the `sphinx/createApp` request and
defines some properties allowing easy access to variables like
`srcdir`, `confdir` etc.
